third_party/blink/web_tests/external/wpt/svg/geometry/svg-baseval-in-display-none-expected.txt

This is a testharness.js-based test.
[FAIL] With em and percentage
  assert_equals: r1.width expected 40 but got 0
Harness: the test ran to completion.
